requirements.txt
pycryptodomex==3.9.9
pywin32==228
